Support more than two macro arguments
Not including `maxsplit=1` meant that only two macro arguments were
ever processed. This new code is a lot simpler and uses a dictionary
comprehension to build the argument lists.

Note that this implementation still fails on some inputs - for example,
any string that contains a comma. However, that's an existing issue, not
affected by this patch. It can only be solved by proper lexing of the
source code, which could be a pretty big task. I'd recommend the use of
`shlex` if you don't want to require a third-party library.